refactor: :truck: rename new charts for brevity
