+++ src/app/insights/components/AreaChart.tsx
