+++ src/app/insights/components/StreamGraph.tsx
